# api/openapi.yaml
openapi: 3.0.3
info:
  title: Nordigen Account Information Services API
  version: 2.0 (v2)
servers:
- url: /
paths:
  /api/v2/accounts/{id}/:
    get:
      description: |-
        Access account metadata.

        Information about the account record, such as the processing status and IBAN.

        Account status is recalculated based on the error count in the latest req.
      operationId: retrieve account metadata
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          description: A UUID string identifying this account.
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountV2'
          description: sample account metadata
        "404":
          content:
            application/json:
              examples:
                NotFoundError:
                  description: Not found error
                  summary: Not found error
                  value:
                    detail: Not found.
                    status_code: 404
                AccountNotFoundError:
                  description: Account not found
                  summary: Account not found error
                  value:
                    summary: 'Account: $ACC_ID not found'
                    detail: ""
                    status_code: 404
              schema:
                additionalProperties: {}
                type: object
      security:
      - tokenAuth: []
      - cookieAuth: []
      tags:
      - accounts
  /api/v2/accounts/{id}/balances/:
    get:
      description: |-
        Access account balances.

        Balances will be returned in Berlin Group PSD2 format.
      operationId: retrieve account balances
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          description: A UUID string identifying this account.
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                balances:
                - balanceAmount:
                    amount: "519.71"
                    currency: string
                  balanceType: string
                  referenceDate: 2021-11-19
                - balanceAmount:
                    amount: -357.79
                    currency: string
                  balanceType: string
                  referenceDate: 2021-11-16
              schema:
                $ref: '#/components/schemas/inline_response_200'
          description: sample account balances
        "400":
          content:
            application/json:
              examples:
                InvalidID:
                  description: Invalid Account ID
                  summary: Invalid ID
                  value:
                    summary: Invalid Account ID
                    detail: $ACCOUNT_ID is not a valid UUID.
                    status_code: 400
              schema:
                additionalProperties: {}
                type: object
        "401":
          content:
            application/json:
              examples:
                AccessExpiredError:
                  description: Couldn't update account balances
                  value:
                    summary: Couldn't update account balances
                    detail: Access has expired or it has been revoked. To restore
                      access reconnect the account.
                    type: AccessExpiredError
                    status_code: 401
                AccountInactiveError:
                  description: Couldn't update account balances
                  value:
                    summary: Couldn't update account balances
                    detail: Account has been deactivated or it no longer exists.
                    type: AccountInactiveError
                    status_code: 401
                FailedAuthentication:
                  description: Couldn't update account balances
                  value:
                    summary: Couldn't update account balances
                    detail: SSN Verification has failed
                    type: FailedAuthentication
                    status_code: 401
              schema:
                additionalProperties: {}
                type: object
        "403":
          content:
            application/json:
              examples:
                AccountEUAScopeError:
                  description: EUA doesn't allow access to account balances.
                  summary: Account EUA scope error
                  value:
                    summary: End User Agreement (EUA) access scope error
                    detail: EUA doesn't allow access to account balances. Check EUA
                      access scope. Or create new EUA with correct access scope
                    status_code: 403
                AccountValidEUAError:
                  description: Account has no valid End User Agreement
                  summary: Account valid EUA error
                  value:
                    summary: No valid End User Agreement
                    error: Account exists but there is no valid End User Agreement
                      permitting you to access it
                    status_code: 403
                AccountAccessForbidden:
                  description: Couldn't update account balances
                  value:
                    summary: Couldn't update account balances
                    detail: Access to account is forbidden. The user might not have
                      the necessary permissions.
                    type: AccountAccessForbidden
                    status_code: 403
              schema:
                additionalProperties: {}
                type: object
        "404":
          content:
            application/json:
              examples:
                NotFoundError:
                  description: Not found error
                  summary: Not found error
                  value:
                    detail: Not found.
                    status_code: 404
                AccountNotFoundError:
                  description: Account not found
                  summary: Account not found error
                  value:
                    summary: 'Account: $ACC_ID not found'
                    detail: ""
                    status_code: 404
              schema:
                additionalProperties: {}
                type: object
        "409":
          content:
            application/json:
              examples:
                AccountSuspendedError:
                  description: Account or its requisition suspended
                  summary: Account suspended error
                  value:
                    summary: Account suspended
                    detail: This account or its requisition was suspended due to numerous
                      errors that occurred while accessing it.
                    status_code: 409
                AccountStateError:
                  description: Account state does not support this operation
                  summary: Account state error
                  value:
                    id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    created: 2021-11-18T23:12:35.151690Z
                    last_accessed: 2021-11-18T23:12:35.151709Z
                    iban: string
                    aspsp_identifier: string
                    status: ERROR
              schema:
                additionalProperties: {}
                type: object
        "429":
          content:
            application/json:
              examples:
                RateLimitError:
                  description: Couldn't update account balances
                  value:
                    summary: Couldn't update account balances
                    detail: Daily request limit set by the Institution has been exceeded.
                    type: RateLimitError
                    status_code: 429
              schema:
                additionalProperties: {}
                type: object
        "500":
          content:
            application/json:
              examples:
                UnknownRequestError:
                  description: Couldn't update account balances
                  value:
                    summary: Couldn't update account balances
                    detail: Request to Institution returned an error
                    type: UnknownRequestError
                    status_code: 500
              schema:
                additionalProperties: {}
                type: object
        "503":
          content:
            application/json:
              examples:
                ServiceError:
                  description: Couldn't update account balances
                  value:
                    summary: Couldn't update account balances
                    detail: Institution service unavailable
                    type: ServiceError
                    status_code: 503
                ConnectionError:
                  description: Couldn't update account balances
                  value:
                    summary: Couldn't update account balances
                    detail: Couldn't connect to Institution
                    type: ConnectionError
                    status_code: 503
              schema:
                additionalProperties: {}
                type: object
      security:
      - tokenAuth: []
      - cookieAuth: []
      tags:
      - accounts
  /api/v2/accounts/{id}/details/:
    get:
      description: |-
        Access account details.

        Account details will be returned in Berlin Group PSD2 format.
      operationId: retrieve account details
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          description: A UUID string identifying this account.
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                account:
                  resourceId: null
                  iban: null
                  currency: null
                  ownerName: null
                  name: null
                  product: null
                  cashAccountType: null
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
          description: sample account details
        "400":
          content:
            application/json:
              examples:
                InvalidID:
                  description: Invalid Account ID
                  summary: Invalid ID
                  value:
                    summary: Invalid Account ID
                    detail: $ACCOUNT_ID is not a valid UUID.
                    status_code: 400
              schema:
                additionalProperties: {}
                type: object
        "401":
          content:
            application/json:
              examples:
                AccessExpiredError:
                  description: Couldn't update account details
                  value:
                    summary: Couldn't update account details
                    detail: Access has expired or it has been revoked. To restore
                      access reconnect the account.
                    type: AccessExpiredError
                    status_code: 401
                AccountInactiveError:
                  description: Couldn't update account details
                  value:
                    summary: Couldn't update account details
                    detail: Account has been deactivated or it no longer exists.
                    type: AccountInactiveError
                    status_code: 401
                FailedAuthentication:
                  description: Couldn't update account details
                  value:
                    summary: Couldn't update account details
                    detail: SSN Verification has failed
                    type: FailedAuthentication
                    status_code: 401
              schema:
                additionalProperties: {}
                type: object
        "403":
          content:
            application/json:
              examples:
                AccountEUAScopeError:
                  description: EUA doesn't allow access to account details.
                  summary: Account EUA scope error
                  value:
                    summary: End User Agreement (EUA) access scope error
                    detail: EUA doesn't allow access to account details. Check EUA
                      access scope. Or create new EUA with correct access scope
                    status_code: 403
                AccountValidEUAError:
                  description: Account has no valid End User Agreement
                  summary: Account valid EUA error
                  value:
                    summary: No valid End User Agreement
                    error: Account exists but there is no valid End User Agreement
                      permitting you to access it
                    status_code: 403
                AccountAccessForbidden:
                  description: Couldn't update account details
                  value:
                    summary: Couldn't update account details
                    detail: Access to account is forbidden. The user might not have
                      the necessary permissions.
                    type: AccountAccessForbidden
                    status_code: 403
              schema:
                additionalProperties: {}
                type: object
        "404":
          content:
            application/json:
              examples:
                NotFoundError:
                  description: Not found error
                  summary: Not found error
                  value:
                    detail: Not found.
                    status_code: 404
                AccountNotFoundError:
                  description: Account not found
                  summary: Account not found error
                  value:
                    summary: 'Account: $ACC_ID not found'
                    detail: ""
                    status_code: 404
              schema:
                additionalProperties: {}
                type: object
        "409":
          content:
            application/json:
              examples:
                AccountSuspendedError:
                  description: Account or its requisition suspended
                  summary: Account suspended error
                  value:
                    summary: Account suspended
                    detail: This account or its requisition was suspended due to numerous
                      errors that occurred while accessing it.
                    status_code: 409
                AccountStateError:
                  description: Account state does not support this operation
                  summary: Account state error
                  value:
                    id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    created: 2021-11-18T23:12:35.151690Z
                    last_accessed: 2021-11-18T23:12:35.151709Z
                    iban: string
                    aspsp_identifier: string
                    status: ERROR
              schema:
                additionalProperties: {}
                type: object
        "429":
          content:
            application/json:
              examples:
                RateLimitError:
                  description: Couldn't update account details
                  value:
                    summary: Couldn't update account details
                    detail: Daily request limit set by the Institution has been exceeded.
                    type: RateLimitError
                    status_code: 429
              schema:
                additionalProperties: {}
                type: object
        "500":
          content:
            application/json:
              examples:
                UnknownRequestError:
                  description: Couldn't update account details
                  value:
                    summary: Couldn't update account details
                    detail: Request to Institution returned an error
                    type: UnknownRequestError
                    status_code: 500
              schema:
                additionalProperties: {}
                type: object
        "503":
          content:
            application/json:
              examples:
                ServiceError:
                  description: Couldn't update account details
                  value:
                    summary: Couldn't update account details
                    detail: Institution service unavailable
                    type: ServiceError
                    status_code: 503
                ConnectionError:
                  description: Couldn't update account details
                  value:
                    summary: Couldn't update account details
                    detail: Couldn't connect to Institution
                    type: ConnectionError
                    status_code: 503
              schema:
                additionalProperties: {}
                type: object
      security:
      - tokenAuth: []
      - cookieAuth: []
      tags:
      - accounts
  /api/v2/accounts/{id}/transactions/:
    get:
      description: |-
        Access account transactions.

        Transactions will be returned in Berlin Group PSD2 format.
      operationId: retrieve account transactions
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          description: A UUID string identifying this account.
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                transactions:
                  booked:
                  - transactionId: string
                    debtorName: string
                    debtorAccount:
                      iban: string
                    transactionAmount:
                      currency: string
                      amount: "996.57"
                    bankTransactionCode: string
                    bookingDate: 2021-11-17T19:48:12.314823
                    valueDate: 2021-11-17T19:48:12.314854
                    remittanceInformationUnstructured: string
                  - transactionId: string
                    transactionAmount:
                      currency: string
                      amount: -438.71
                    bankTransactionCode: string
                    bookingDate: 2021-11-14T19:48:12.314869
                    valueDate: 2021-11-16T19:48:12.314876
                    remittanceInformationUnstructured: string
                  pending:
                  - transactionAmount:
                      currency: string
                      amount: -135.56
                    valueDate: 2021-11-17T19:48:12.314890
                    remittanceInformationUnstructured: string
                  - transactionAmount:
                      currency: string
                      amount: -135.56
                    valueDate: 2021-11-17T19:48:12.314890
                    remittanceInformationUnstructured: string
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
          description: sample transaction data
        "400":
          content:
            application/json:
              examples:
                InvalidID:
                  description: Invalid Account ID
                  summary: Invalid ID
                  value:
                    summary: Invalid Account ID
                    detail: $ACCOUNT_ID is not a valid UUID.
                    status_code: 400
              schema:
                additionalProperties: {}
                type: object
        "401":
          content:
            application/json:
              examples:
                AccessExpiredError:
                  description: Couldn't update account transactions
                  value:
                    summary: Couldn't update account transactions
                    detail: Access has expired or it has been revoked. To restore
                      access reconnect the account.
                    type: AccessExpiredError
                    status_code: 401
                AccountInactiveError:
                  description: Couldn't update account transactions
                  value:
                    summary: Couldn't update account transactions
                    detail: Account has been deactivated or it no longer exists.
                    type: AccountInactiveError
                    status_code: 401
                FailedAuthentication:
                  description: Couldn't update account transactions
                  value:
                    summary: Couldn't update account transactions
                    detail: SSN Verification has failed
                    type: FailedAuthentication
                    status_code: 401
              schema:
                additionalProperties: {}
                type: object
        "403":
          content:
            application/json:
              examples:
                AccountEUAScopeError:
                  description: EUA doesn't allow access to account transactions.
                  summary: Account EUA scope error
                  value:
                    summary: End User Agreement (EUA) access scope error
                    detail: EUA doesn't allow access to account transactions. Check
                      EUA access scope. Or create new EUA with correct access scope
                    status_code: 403
                AccountValidEUAError:
                  description: Account has no valid End User Agreement
                  summary: Account valid EUA error
                  value:
                    summary: No valid End User Agreement
                    error: Account exists but there is no valid End User Agreement
                      permitting you to access it
                    status_code: 403
                AccountAccessForbidden:
                  description: Couldn't update account transactions
                  value:
                    summary: Couldn't update account transactions
                    detail: Access to account is forbidden. The user might not have
                      the necessary permissions.
                    type: AccountAccessForbidden
                    status_code: 403
              schema:
                additionalProperties: {}
                type: object
        "404":
          content:
            application/json:
              examples:
                NotFoundError:
                  description: Not found error
                  summary: Not found error
                  value:
                    detail: Not found.
                    status_code: 404
                AccountNotFoundError:
                  description: Account not found
                  summary: Account not found error
                  value:
                    summary: 'Account: $ACC_ID not found'
                    detail: ""
                    status_code: 404
              schema:
                additionalProperties: {}
                type: object
        "409":
          content:
            application/json:
              examples:
                AccountSuspendedError:
                  description: Account or its requisition suspended
                  summary: Account suspended error
                  value:
                    summary: Account suspended
                    detail: This account or its requisition was suspended due to numerous
                      errors that occurred while accessing it.
                    status_code: 409
                AccountStateError:
                  description: Account state does not support this operation
                  summary: Account state error
                  value:
                    id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    created: 2021-11-18T23:12:35.151690Z
                    last_accessed: 2021-11-18T23:12:35.151709Z
                    iban: string
                    aspsp_identifier: string
                    status: ERROR
              schema:
                additionalProperties: {}
                type: object
        "429":
          content:
            application/json:
              examples:
                RateLimitError:
                  description: Couldn't update account transactions
                  value:
                    summary: Couldn't update account transactions
                    detail: Daily request limit set by the Institution has been exceeded.
                    type: RateLimitError
                    status_code: 429
              schema:
                additionalProperties: {}
                type: object
        "500":
          content:
            application/json:
              examples:
                UnknownRequestError:
                  description: Couldn't update account transactions
                  value:
                    summary: Couldn't update account transactions
                    detail: Request to Institution returned an error
                    type: UnknownRequestError
                    status_code: 500
              schema:
                additionalProperties: {}
                type: object
        "503":
          content:
            application/json:
              examples:
                ServiceError:
                  description: Couldn't update account transactions
                  value:
                    summary: Couldn't update account transactions
                    detail: Institution service unavailable
                    type: ServiceError
                    status_code: 503
                ConnectionError:
                  description: Couldn't update account transactions
                  value:
                    summary: Couldn't update account transactions
                    detail: Couldn't connect to Institution
                    type: ConnectionError
                    status_code: 503
              schema:
                additionalProperties: {}
                type: object
      security:
      - tokenAuth: []
      - cookieAuth: []
      tags:
      - accounts
...
components:
  schemas:
    AccountV2:
      description: AccountV2Serializer.
      example:
        last_accessed: 2000-01-23T04:56:07.000+00:00
        created: 2000-01-23T04:56:07.000+00:00
        iban: iban
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        institution_id: institution_id
        status: ""
      properties:
        id:
          description: The ID of this Account, used to refer to this account in other
            API calls.
          format: uuid
          readOnly: true
          title: Account ID
          type: string
        created:
          description: The date & time at which the account object was created.
          format: date-time
          readOnly: true
          title: Created date/time
          type: string
        last_accessed:
          description: The date & time at which the account object was last accessed.
          format: date-time
          nullable: true
          readOnly: true
          title: Last accessed date/time
          type: string
        iban:
          description: The Account IBAN
          readOnly: true
          type: string
        institution_id:
          description: The ASPSP associated with this account.
          readOnly: true
          title: ASPSP ID
          type: string
        status:
          allOf:
          - $ref: '#/components/schemas/AccountV2StatusEnum'
          description: The processing status of this account.
          readOnly: true
          title: Account Status
      type: object
    AccountV2StatusEnum:
      enum:
      - '{"DISCOVERED":"User has successfully authenticated and account is discovered"}'
      - '{"PROCESSING":"Account is being processed by the Institution"}'
      - '{"ERROR":"An error was encountered when processing account"}'
      - '{"EXPIRED":"Access to account has expired as set in End User Agreement"}'
      - '{"READY":"Account has been successfully processed"}'
      - '{"SUSPENDED":"Account has been suspended (more than 10 consecutive failed
        attempts to access the account)"}'
      type: string
...
    inline_response_200:
      example:
        balances:
        - '{}'
        - '{}'
      properties:
        balances:
          items:
            properties:
              balanceType:
                title: Balance Type
                type: string
              referenceDate:
                title: Reference Date
                type: string
              balanceAmount:
                properties:
                  amount:
                    title: Amount
                    type: string
                  currency:
                    title: Currency
                    type: string
                type: object
          type: array
      type: object
    inline_response_200_1_account:
      example:
        cashAccountType: cashAccountType
        resourceId: resourceId
        product: product
        ownerName: ownerName
        iban: iban
        name: name
        currency: currency
      properties:
        resourceId:
          title: ResourceID
          type: string
        iban:
          type: string
        currency:
          type: string
        ownerName:
          type: string
        name:
          type: string
        product:
          type: string
        cashAccountType:
          type: string
      type: object
    inline_response_200_1:
      example:
        account:
          cashAccountType: cashAccountType
          resourceId: resourceId
          product: product
          ownerName: ownerName
          iban: iban
          name: name
          currency: currency
      properties:
        account:
          $ref: '#/components/schemas/inline_response_200_1_account'
      type: object
...
    inline_response_200_2:
      example:
        transactions:
          booked:
          - '{}'
          - '{}'
          pending:
          - '{}'
          - '{}'
      properties:
        transactions:
          $ref: '#/components/schemas/inline_response_200_2_transactions'
      type: object
  securitySchemes:
    cookieAuth:
      in: cookie
      name: Session
      type: apiKey
    jwtAuth:
      bearerFormat: Bearer
      scheme: bearer
      type: http
    tokenAuth:
      description: Token-based authentication with required prefix "Token"
      in: header
      name: Authorization
      type: apiKey